+++ extension/app/panel-popup/open.js
@@ -1,9 +1,8 @@
 import { openPanelInActionPopup } from "./open-action-popup.js";
-import { openPanelInTab } from "../panel-tab/open.js";
 import { PANEL_PAGE_CONFIG } from "./constants.js";
 
 export function openPanelInActionPopup2(panelTab, target) {
-  openPanelInActionPopup(PANEL_PAGE_CONFIG, panelTab, target, openPanelInTab);
+  openPanelInActionPopup(PANEL_PAGE_CONFIG, panelTab, target);
 }
 export function openPanelFromSender(panelTab, senderTab) {
   openPanelInActionPopup2(panelTab, {